#13663: Relocate sharding_utilities
